drivers: can: rework zcan_frame and zcan_filter
Reordering of the struct elements to match the Linux format.
The __packed() is not necessary anymore.
std_id and ext_id is merged to id in the frame and filter.
Additionally, the frames are ready for CAN-FD and the data
field can hold either the data or a pointer to the data.

Signed-off-by: Alexander Wachter <alexander@wachter.cloud>